front: move ManageTimetableItem folder in views/Scenario
