+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/ConstraintDistributionSwitch.tsx
