+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/ElectricalProfiles.tsx
