+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/ManageTimetableItemMap/OperationalPointPopupDetails.tsx
