+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/ManageTimetableItemModal.tsx
@@ -1,13 +1,13 @@
 import { ChevronLeft, ChevronRight } from '@osrd-project/ui-icons';
 
 import { ManageTimetableItemContextProvider } from 'applications/operationalStudies/hooks/useManageTimetableItemContext';
+
+import ImportTimetableItem from '../ImportTimetableItem';
+import ManageTimetableItem from './ManageTimetableItem';
 import ManageTimetableItemLeftPanel, {
   type ManageTimetableItemLeftPanelProps,
-} from 'modules/timetableItem/components/ManageTimetableItem/ManageTimetableItemLeftPanel';
-
-import ImportTimetableItem from './ImportTimetableItem';
-import ManageTimetableItem from './ManageTimetableItem';
-import { MANAGE_TIMETABLE_ITEM_TYPES } from '../consts';
+} from './ManageTimetableItemLeftPanel';
+import { MANAGE_TIMETABLE_ITEM_TYPES } from '../../consts';
 
 type ManageTimetableItemModalProps = ManageTimetableItemLeftPanelProps & {
   setCollapsedTimetableEdit: () => void;
