+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/PacedTrainSettings/PacedTrainSettings.tsx
