+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/SimulationSettings.tsx
