+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/TrainSettings.tsx
