+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/consts.ts
