+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/helpers/__tests__/formatSchedule.spec.ts
