+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/helpers/buildPacedTrainException.ts
