+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/helpers/formatMargin.ts
@@ -1,5 +1,6 @@
-import type { Margin } from 'modules/timetableItem/components/ManageTimetableItem/types';
 import type { PathStep } from 'reducers/osrdconf/types';
+
+import type { Margin } from '../types';
 
 const formatMargin = (pathSteps: PathStep[]): Margin | undefined => {
   const boundaries: string[] = [];
