+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/helpers/formatSchedule.ts
