+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/helpers/formatTimetableItemPayload.ts
@@ -3,6 +3,7 @@ import { v4 as uuidV4 } from 'uuid';
 
 import type { PacedTrain, TrainSchedule } from 'common/api/osrdEditoastApi';
 import getStepLocation from 'modules/pathfinding/helpers/getStepLocation';
+import type { PacedTrainWithDetails } from 'modules/timetableItem/components/Timetable/types';
 import { findExceptionWithOccurrenceId } from 'modules/timetableItem/helpers/pacedTrain';
 import type {
   TimetableItemToEditData,
@@ -23,7 +24,6 @@ import {
 } from './buildPacedTrainException';
 import formatMargin from './formatMargin';
 import formatSchedule from './formatSchedule';
-import type { PacedTrainWithDetails } from '../../Timetable/types';
 
 export function formatTimetableItemPayload(
   osrdconf: OperationalStudiesConfState,
