+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/index.ts
@@ -1,0 +1,3 @@
+import ManageTimetableItemModal from './ManageTimetableItemModal';
+
+export default ManageTimetableItemModal;
