+++ front/src/applications/operationalStudies/views/Scenario/components/ManageTimetableItem/types.ts
