+++ front/src/applications/operationalStudies/views/Scenario/components/ScenarioContent.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next';
 import { useScenarioContext } from 'applications/operationalStudies/hooks/useScenarioContext';
 import useScenarioData from 'applications/operationalStudies/hooks/useScenarioData';
 import type { Board } from 'applications/operationalStudies/types';
-import ManageTimetableItemModal from 'applications/operationalStudies/views/Scenario/components/ManageTimetableItemModal';
+import ManageTimetableItemModal from 'applications/operationalStudies/views/Scenario/components/ManageTimetableItem';
 import SimulationResults from 'applications/operationalStudies/views/Scenario/components/SimulationResults';
 import type { Conflict } from 'common/api/osrdEditoastApi';
 import { Loader } from 'common/Loaders';
