+++ front/src/modules/simulationResult/components/SimulationResultsMap/SimulationResultsMap.tsx
@@ -7,6 +7,7 @@ import type { MapRef } from 'react-map-gl/maplibre';
 import captureMap from 'applications/operationalStudies/helpers/captureMap';
 import { useScenarioContext } from 'applications/operationalStudies/hooks/useScenarioContext';
 import type { PathPropertiesFormatted } from 'applications/operationalStudies/types';
+import { MARKER_TYPE } from 'applications/operationalStudies/views/Scenario/components/ManageTimetableItem/ManageTimetableItemMap/ItineraryMarkers';
 import type { PathfindingResultSuccess } from 'common/api/osrdEditoastApi';
 import BaseMap from 'common/Map/BaseMap';
 import MapButtons from 'common/Map/Buttons/MapButtons';
@@ -16,7 +17,6 @@ import { useInfraID } from 'common/osrdContext';
 import { LAYER_GROUPS_ORDER, LAYERS } from 'config/layerOrder';
 import getPointOnPathCoordinates from 'modules/pathfinding/helpers/getPointOnPathCoordinates';
 import getTrackLengthCumulativeSums from 'modules/pathfinding/helpers/getTrackLengthCumulativeSums';
-import { MARKER_TYPE } from 'modules/timetableItem/components/ManageTimetableItem/ManageTimetableItemMap/ItineraryMarkers';
 import { useMapSettings, useMapSettingsActions } from 'reducers/globalMap';
 import type { Viewport } from 'reducers/globalMap/types';
 import { useAppDispatch } from 'store';
