--- front/src/modules/timetableItem/components/ManageTimetableItem/index.ts
@@ -1,2 +1,0 @@
-export { default as TrainSettings } from 'modules/timetableItem/components/ManageTimetableItem/TrainSettings';
-export { default as ElectricalProfiles } from 'modules/timetableItem/components/ManageTimetableItem/ElectricalProfiles';
